backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
duckdb>=0.10.0
pandas>=2.2.0
numpy>=1.26.0
akshare>=1.14.0
httpx>=0.27.0
TA-Lib>=0.4.28
apscheduler>=3.10.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
openai>=1.30.0
python-dotenv>=1.0.0
